Refactored DriverScreen fully ready - missing connection with server .
